consolidate image conversion classes
