+++ src/main/java/com/marginallyclever/makelangelo/makeart/imageconverter/ImageConverter.java
@@ -7,7 +7,6 @@ import com.marginallyclever.convenience.Clipper2D;
 import com.marginallyclever.convenience.MathHelper;
 import com.marginallyclever.convenience.Point2D;
 import com.marginallyclever.makelangelo.makeart.TransformedImage;
-import com.marginallyclever.makelangelo.makeart.io.image.ImageConverterThread;
 import com.marginallyclever.makelangelo.paper.Paper;
 import com.marginallyclever.makelangelo.turtle.Turtle;
