+++ src/main/java/com/marginallyclever/makelangelo/makeart/io/OpenFileChooser.java
@@ -2,7 +2,7 @@ package com.marginallyclever.makelangelo.makeart.io;
 
 import com.marginallyclever.convenience.FileAccess;
 import com.marginallyclever.makelangelo.Translator;
-import com.marginallyclever.makelangelo.makeart.io.image.SelectImageConverterPanel;
+import com.marginallyclever.makelangelo.makeart.imageconverter.SelectImageConverterPanel;
 import com.marginallyclever.makelangelo.makeart.io.vector.TurtleFactory;
 import com.marginallyclever.util.PreferencesHelper;
 import org.slf4j.Logger;
